refactor: move runtime modules out of src/types
- SWRResponseTypes -> hooks/utils/swr, next to the hook factories that return it
- taxCodeComboBoxOption -> utils/features/bankTransactions, beside categorizationOption
- delete utility/branded.ts, which had no consumers and ran Symbol.for() on import
