+++ AGENTS.md
@@ -167,8 +167,8 @@ Reach for these before writing your own:
 - `readonly`/`ReadonlyArray` for data you don't own; `asMutable` (`@utils/shared/array/asMutable`) at the
   boundary of an API that demands a mutable array.
 - Shared utility types live in `src/types/utility/**`: `OneOf` (exclusive unions),
-  `EnumWithUnknownValues` (open string enums), branded `EmailAddress`/`PhoneNumber`,
-  `pagination`, `promises`. Check there before writing a new type-level helper.
+  `EnumWithUnknownValues` (open string enums), `pagination`, `promises`, `table`. Check there
+  before writing a new type-level helper.
 - `src/types/**` is for internal-only types with no wire format. Anything the API sends or
   receives is a schema.
 - Not yet enabled in `tsconfig.json` but worth honoring: `isolatedModules`, `verbatimModuleSyntax`, `noUncheckedIndexedAccess`.
