+++ src/components/features/bankTransactions/BankTransactionsBulkActions/BankTransactionsCategorizeAllModal.tsx
@@ -2,8 +2,8 @@ import { useCallback, useId, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { isApiCategorizationAsOption, isCategoryAsOption } from '@internal-types/bankTransactionCategoryComboBoxOption'
-import type { TaxCodeComboBoxOption } from '@internal-types/taxCodeComboBoxOption'
 import { canCategoryHaveTaxCode, resolveCategoryTaxCode } from '@utils/features/bankTransactions/taxCode'
+import type { TaxCodeComboBoxOption } from '@utils/features/bankTransactions/taxCodeComboBoxOption'
 import { tPlural } from '@utils/shared/i18n/plural'
 import { useBulkCategorizeBankTransactions } from '@hooks/features/bankTransactions/useBulkBankTransactionMutations'
 import { useTaxCodeOptions } from '@hooks/features/bankTransactions/useTaxCodeOptions'
