+++ src/components/features/bankTransactions/BankTransactionsMobileCategorySelection/BankTransactionsMobileCategorySelection.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next'
 
 import { isPlaceholderAsOption } from '@internal-types/bankTransactionCategoryComboBoxOption'
 import { type BankTransaction } from '@internal-types/bankTransactions'
-import type { TaxCodeComboBoxOption } from '@internal-types/taxCodeComboBoxOption'
 import { canCategoryHaveTaxCode } from '@utils/features/bankTransactions/taxCode'
+import type { TaxCodeComboBoxOption } from '@utils/features/bankTransactions/taxCodeComboBoxOption'
 import { useGetBankTransactionCategorizationWithDefault } from '@hooks/features/bankTransactions/useGetBankTransactionCategorizationWithDefault'
 import { useTaxCodeOptions } from '@hooks/features/bankTransactions/useTaxCodeOptions'
 import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
