+++ src/hooks/api/businesses/[business-id]/get.ts
@@ -1,11 +1,11 @@
 import useSWR from 'swr'
 
-import { SWRQueryResult } from '@internal-types/swr/SWRResponseTypes'
 import { BusinessResponseSchema } from '@schemas/business/business'
 import { get } from '@utils/shared/api/authenticatedHttp'
 import { createBuildKey } from '@utils/shared/swr/createBuildKey'
 import { createKeyedFetcher } from '@utils/shared/swr/createKeyedFetcher'
 import { useAuth } from '@hooks/utils/auth/useAuth'
+import { SWRQueryResult } from '@hooks/utils/swr/SWRResponseTypes'
 
 export const BUSINESS_TAG_KEY = '#business'
