+++ src/hooks/utils/swr/createQueryHook.ts
@@ -1,9 +1,9 @@
 import { type Schema } from 'effect'
 import useSWR, { type SWRConfiguration } from 'swr'
 
-import { SWRQueryResult } from '@internal-types/swr/SWRResponseTypes'
 import { createBuildKey } from '@utils/shared/swr/createBuildKey'
 import { type AuthenticatedRequest, createKeyedFetcher, type SWRKeyContext } from '@utils/shared/swr/createKeyedFetcher'
+import { SWRQueryResult } from '@hooks/utils/swr/SWRResponseTypes'
 import { useBuildKeyInputs } from '@hooks/utils/swr/useBuildKeyInputs'
 
 type BusinessScopedParams = { businessId: string }
